Settings: Slim heads up customizations for LP (2/2)
Basically it follows the same idea like we had on SlimKat....see here
https://github.com/SlimRoms/frameworks_base/commit/323ed77c5e5bd6d421beae1f8118eb580be5ecf1

That said there are some differences in this more or less completely
recoded feature.

- Moved it into the app policy files
- Enable/Disable on every app (no system apps)  either longpress a
notification and an button appears / or in notification app settings
- For now only snooze and heads up timeout configurable....other stuff
will follow later.
- fix up some AOSP code for it as well fix some UX inconsitence.
- left swipe dismiss, right swipe put into background, swipe down opens
notification drawer and moves the notification into it
(aosp default behaviour is a big UX disaster....swipe up dismisses it
completely and swipe down open notification drawer but heads up
stays....first
   happened to me 10000 times accidentally which is really bad and
second does not make sense. Hence this change of the default behaviour)

Well and a lot smaller things as usual.

Change-Id: I96c3259ad705b1776c58d2119410cd97d5f6d928