Settings: report stats upload immediately
We should report the upload status immediately to the job service and
not let the executor leave doInBackground() even after the job
successfully finished, and report a potential failure in onPostExecut().
Otherwise if conditions are no longer met, and the job has finished, it can
result in the job reporting as failed, even though it has succeeded.

Ref: CYNGNOS-714

Change-Id: I5e06ae9dbca5c1412641eed51acded7c67522ded
Signed-off-by: Roman Birg <roman@cyngn.com>