[1/2] Settings: Add Button to Recents to open App as Floating
I considered the current implementations to launch an App in floating
mode as to uncomfortable, destroying all the workflow MultiWindow could
bring. This commits adds an Button to each TaskView which launches
the Task in a Floating Window. As maybe not everybody wants that, I
implemented a way to turn it off in Settings.

Original Commit for CarbonROM. Powered by the mighty Flying Spagetti Monster.
~;;o;°°;;o;~

Change-Id: I5849a9b25bb0e19be91b3e98b48b0c95934383e2